Adding public/private route guard
Adding guards to the application. Using the public guard to redirect users to the todos page if they try to navigate to sign in, sign up, or forgot password while already signed in. Using the private guard to redirect users to the sign in page if they try to navigate to todos, files, and account while not signed in.